Fix some bugs, add theming system, and improve state resets
The state should now properly reset when something causes the UI to show a reset (like a reset of the ink story object). Also added a theming system for stories to theme the UI through variables. Tutorial story added to show the game without going into the incomplete fantasy story.